+++ services/ui-src/src/components/reports/ReportPage.test.tsx
@@ -1,5 +1,4 @@
 import { render, screen } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
 import { axe } from "jest-axe";
 // components
 import { ReportContext, ReportPage } from "components";
@@ -89,16 +88,6 @@ describe("Test ReportPage functionality", () => {
   test("ReportPage navigates to dashboard if no id", () => {
     render(ReportPageComponent_WithoutReport);
     expect(mockUseNavigate).toHaveBeenCalledWith("/mcpar");
-  });
-
-  test("ReportPage updates report field data on successful fill", async () => {
-    const result = render(ReportPageComponent_StaticPage);
-    const form = result.container;
-    const mockField = form.querySelector("[name='mock-text-field']")!;
-    await userEvent.type(mockField, "mock input");
-    const submitButton = form.querySelector("[type='submit']")!;
-    await userEvent.click(submitButton);
-    expect(mockReportContext.updateReport).toHaveBeenCalledTimes(1);
   });
 });
